app/src/main/java/com/fetch/androidtakehome/Network/APIClient.java
package com.fetch.androidtakehome.Network;

import retrofit2.Retrofit;
import retrofit2.converter.gson.GsonConverterFactory;

/**
 * Responsible for creating and managing a singleton instance of Retrofit.
 * It provides a configured Retrofit instance for making API calls.
 */
public class APIClient {

    private static final String BASE_URL = "https://fetch-hiring.s3.amazonaws.com/";
    private static Retrofit retrofit;

    public static Retrofit getRetrofitInstance() {
        if (retrofit == null) {
            // Initialize the Retrofit instance if it doesn't exist
            retrofit = new Retrofit.Builder()
                    .baseUrl(BASE_URL)
                    .addConverterFactory(GsonConverterFactory.create())
                    .build();
        }
        return retrofit;
    }
}